fix: u32 auction rate bump

+++ src/domains/auction-details/hasher.ts
@@ -11,7 +11,7 @@ const types = {
                 fields: [
                     {name: 'startTime', type: 'u32'},
                     {name: 'duration', type: 'u32'},
-                    {name: 'initialRateBump', type: 'u16'},
+                    {name: 'initialRateBump', type: 'u32'},
                     {
                         name: 'pointsAndTimeDeltas',
                         type: {vec: {defined: {name: 'pointAndTimeDelta'}}}
@@ -24,7 +24,7 @@ const types = {
             type: {
                 kind: 'struct',
                 fields: [
-                    {name: 'rateBump', type: 'u16'},
+                    {name: 'rateBump', type: 'u32'},
                     {name: 'timeDelta', type: 'u16'}
                 ]
             }
